+++ src/package.json
@@ -8,22 +8,25 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/vector-icons": "^14.1.0",
+    "@react-native-async-storage/async-storage": "2.1.2",
+    "@react-navigation/drawer": "*",
+    "@react-navigation/native": "*",
+    "@react-navigation/stack": "*",
     "expo": "~53.0.9",
+    "expo-location": "~18.1.5",
+    "expo-secure-store": "~14.2.3",
     "expo-status-bar": "~2.2.3",
     "react": "19.0.0",
     "react-native": "0.79.2",
-    "expo-location": "~18.1.5",
-    "expo-secure-store": "~14.2.3",
-    "@expo/vector-icons": "^14.1.0",
-    "react-native-paper": "4.9.2",
-    "react-native-screens": "~4.10.0",
-    "@react-navigation/stack": "*",
-    "react-native-reanimated": "~3.17.4",
-    "@react-navigation/drawer": "*",
-    "@react-navigation/native": "*",
     "react-native-gesture-handler": "~2.24.0",
+    "react-native-paper": "4.9.2",
+    "react-native-reanimated": "~3.17.4",
     "react-native-safe-area-context": "5.4.0",
-    "@react-native-async-storage/async-storage": "2.1.2"
+    "react-native-screens": "~4.10.0",
+    "react-dom": "19.0.0",
+    "react-native-web": "^0.20.0",
+    "@expo/metro-runtime": "~5.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
